Reduce duplication on blueprint/controllers
