+++ requirements.txt
@@ -1,3 +1,4 @@
+ckantoolkit
 GeoAlchemy>=0.6
 GeoAlchemy2==0.5.0
 Shapely>=1.2.13
